feat: add location field to the job schema
This commit introduces a new field called "country" to the job schema, replacing the "location" field. The "country" field will be used for filtering and searching jobs based on geographic location. The change addresses a need for more precise location information and simplifies the search functionality by using a single country field.